API Insert be kinda pog ngl
Ok so anyways imma leave this message here but i doubt anyone would read it. So what i did is i basically inserted all the countries data into the table and made sure all the empty spaces will be NULL in the database. Have a great day to whoever read this part